+++ server/app/service/word_service.py
@@ -4,7 +4,7 @@ import json
 from app.module.socket_module import socketio
 from app.module.logging_module import logger
 from app.module.dictionary_module import ignore_list, ignore_list_file
-from app.service import anki_service
+from app.service.anki import anki_service
 from app.repository.word.word_repository import word_repository
